Skip info cache update in case instance is deleted
The network api is sometimes called for a deleted instance causing
info cache update to fail. An example of this is the compute manager
periodic task cleanup_running_deleted_instances() that tries to
shutdown deleted instances still running on a host.

Add a check in network_api.update_instance_cache_with_nw_info() that
will skip info cache update if the instance is deleted.

Change-Id: Ibcf2db92275c9860409793d469ed688e7024120b
Closes-Bug: #1282858