Supernova - remove non-Docker Nova code
Tengi era of the emperor Go-Reizei, second year, fourth lunar month,
after the middle period of ten days. At chou, a guest star appeared
in the degrees of the moon mansions of Zuixi and Shen. It has been
viewed in the direction of the East and has emerged from the
Tianguan star. It was as big as Jupiter. - Fujiwara no Teika

See reasoning behind removing this code here:
  https://medium.com/p/71fefeffd781

Change-Id: I29ac66c6754d9f96c2a6b216e4a215d87ff271cb